moved gui_common to core
