+++ core/gui_common.py
