+++ tests/test_gui_common.py
@@ -80,47 +80,47 @@ class TestColorConstants(unittest.TestCase):
 
     def test_color_primary_exists(self):
         """Given gui_common module, Then COLOR_PRIMARY should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_PRIMARY'))
 
     def test_color_success_exists(self):
         """Given gui_common module, Then COLOR_SUCCESS should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_SUCCESS'))
 
     def test_color_error_exists(self):
         """Given gui_common module, Then COLOR_ERROR should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_ERROR'))
 
     def test_color_warning_exists(self):
         """Given gui_common module, Then COLOR_WARNING should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_WARNING'))
 
     def test_color_bg_exists(self):
         """Given gui_common module, Then COLOR_BG should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_BG'))
 
     def test_color_bg_light_exists(self):
         """Given gui_common module, Then COLOR_BG_LIGHT should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_BG_LIGHT'))
 
     def test_color_secundary_exists(self):
         """Given gui_common module, Then COLOR_SECUNDARY should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_SECUNDARY'))
 
     def test_color_disconnected_exists(self):
         """Given gui_common module, Then COLOR_DISCONNECTED should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'COLOR_DISCONNECTED'))
 
     def test_color_disconnected_is_gray(self):
         """Given COLOR_DISCONNECTED, Then it should be a gray color tuple."""
-        import gui_common
+        from core import gui_common
         self.assertEqual(gui_common.COLOR_DISCONNECTED, (0.7, 0.7, 0.7, 1))
 
 
@@ -133,24 +133,24 @@ class TestConnectionState(unittest.TestCase):
 
     def test_connection_state_exists(self):
         """Given gui_common module, Then ConnectionState should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'ConnectionState'))
 
     def test_connection_state_has_text_attribute(self):
         """Given ConnectionState, Then it should have a text attribute."""
-        import gui_common
+        from core import gui_common
         state = gui_common.ConnectionState(text='Test', color=(1, 1, 1, 1))
         self.assertEqual(state.text, 'Test')
 
     def test_connection_state_has_color_attribute(self):
         """Given ConnectionState, Then it should have a color attribute."""
-        import gui_common
+        from core import gui_common
         state = gui_common.ConnectionState(text='Test', color=(0.5, 0.5, 0.5, 1))
         self.assertEqual(state.color, (0.5, 0.5, 0.5, 1))
 
     def test_connection_state_is_frozen(self):
         """Given ConnectionState, Then it should be immutable (frozen)."""
-        import gui_common
+        from core import gui_common
         state = gui_common.ConnectionState(text='Test', color=(1, 1, 1, 1))
         with self.assertRaises(AttributeError):
             state.text = 'New Text'
@@ -165,68 +165,68 @@ class TestGetLogColor(unittest.TestCase):
 
     def test_get_log_color_exists(self):
         """Given gui_common module, Then get_log_color should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'get_log_color'))
         self.assertTrue(callable(gui_common.get_log_color))
 
     def test_get_log_color_returns_error_for_error_level(self):
         """Given ERROR log level, Then get_log_color should return COLOR_ERROR."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.ERROR, "Some error")
         self.assertEqual(color, gui_common.COLOR_ERROR)
 
     def test_get_log_color_returns_error_for_critical_level(self):
         """Given CRITICAL log level, Then get_log_color should return COLOR_ERROR."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.CRITICAL, "Critical error")
         self.assertEqual(color, gui_common.COLOR_ERROR)
 
     def test_get_log_color_returns_warning_for_warning_level(self):
         """Given WARNING log level, Then get_log_color should return COLOR_WARNING."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.WARNING, "Some warning")
         self.assertEqual(color, gui_common.COLOR_WARNING)
 
     def test_get_log_color_returns_success_for_success_keyword(self):
         """Given INFO level with 'success' in message, Then return COLOR_SUCCESS."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.INFO, "Operation success")
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
     def test_get_log_color_returns_success_for_txid_keyword(self):
         """Given INFO level with 'txid' in message, Then return COLOR_SUCCESS."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.INFO, "TXID: abc123")
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
     def test_get_log_color_returns_success_for_broadcast_keyword(self):
         """Given INFO level with 'broadcast' in message, Then return COLOR_SUCCESS."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.INFO, "Transaction broadcast")
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
     def test_get_log_color_returns_success_for_ack_keyword(self):
         """Given INFO level with 'ack' in message, Then return COLOR_SUCCESS."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.INFO, "Received ACK")
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
     def test_get_log_color_returns_none_for_normal_info(self):
         """Given INFO level with normal message, Then return None."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.INFO, "Normal info message")
         self.assertIsNone(color)
 
     def test_get_log_color_with_custom_success_keywords(self):
         """Given custom success keywords, Then use those for matching."""
-        import gui_common
+        from core import gui_common
         # Default keywords don't include 'complete'
         color = gui_common.get_log_color(logging.INFO, "Task complete", success_keywords=['complete'])
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
     def test_get_log_color_case_insensitive(self):
         """Given message with mixed case, Then matching should be case-insensitive."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_log_color(logging.INFO, "SUCCESS message")
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
@@ -240,43 +240,43 @@ class TestGetPrintColor(unittest.TestCase):
 
     def test_get_print_color_exists(self):
         """Given gui_common module, Then get_print_color should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'get_print_color'))
         self.assertTrue(callable(gui_common.get_print_color))
 
     def test_get_print_color_returns_error_for_error_keyword(self):
         """Given message with 'error', Then return COLOR_ERROR."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_print_color("Error occurred")
         self.assertEqual(color, gui_common.COLOR_ERROR)
 
     def test_get_print_color_returns_error_for_failed_keyword(self):
         """Given message with 'failed', Then return COLOR_ERROR."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_print_color("Connection failed")
         self.assertEqual(color, gui_common.COLOR_ERROR)
 
     def test_get_print_color_returns_error_for_abort_keyword(self):
         """Given message with 'abort', Then return COLOR_ERROR."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_print_color("User abort")
         self.assertEqual(color, gui_common.COLOR_ERROR)
 
     def test_get_print_color_returns_success_for_success_keyword(self):
         """Given message with 'success', Then return COLOR_SUCCESS."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_print_color("Operation success")
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
     def test_get_print_color_returns_success_for_txid_keyword(self):
         """Given message with 'txid', Then return COLOR_SUCCESS."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_print_color("TXID: abc123def456")
         self.assertEqual(color, gui_common.COLOR_SUCCESS)
 
     def test_get_print_color_returns_none_for_normal_message(self):
         """Given normal message, Then return None."""
-        import gui_common
+        from core import gui_common
         color = gui_common.get_print_color("Normal message")
         self.assertIsNone(color)
 
@@ -290,7 +290,7 @@ class TestStatusLog(unittest.TestCase):
 
     def test_status_log_class_exists(self):
         """Given gui_common module, Then StatusLog class should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'StatusLog'))
 
 
@@ -303,31 +303,31 @@ class TestWidgetFactories(unittest.TestCase):
 
     def test_create_separator_exists(self):
         """Given gui_common module, Then create_separator should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'create_separator'))
         self.assertTrue(callable(gui_common.create_separator))
 
     def test_create_title_exists(self):
         """Given gui_common module, Then create_title should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'create_title'))
         self.assertTrue(callable(gui_common.create_title))
 
     def test_create_section_label_exists(self):
         """Given gui_common module, Then create_section_label should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'create_section_label'))
         self.assertTrue(callable(gui_common.create_section_label))
 
     def test_create_clear_button_exists(self):
         """Given gui_common module, Then create_clear_button should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'create_clear_button'))
         self.assertTrue(callable(gui_common.create_clear_button))
 
     def test_create_action_button_exists(self):
         """Given gui_common module, Then create_action_button should be defined."""
-        import gui_common
+        from core import gui_common
         self.assertTrue(hasattr(gui_common, 'create_action_button'))
         self.assertTrue(callable(gui_common.create_action_button))
